+++ HISTORY.md
@@ -1,6 +1,7 @@
 unreleased
 ==========
 
+  * Accept string for `maxAge` (converted by `ms`)
   * Use `etag` to generate `ETag` header
 
 2.0.1 / 2014-06-05